pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "GlocalText"
version = "3.0.0"
authors = [{ name = "OG-Open-Source" }]
description = "A config-driven script to batch translate files using various providers."
requires-python = ">=3.8"
classifiers = [
  "Programming Language :: Python :: 3",
  "Operating System :: OS Independent",
]
dependencies = ["deep-translator", "google-genai", "PyYAML", "regex"]

[project.scripts]
glocaltext = "glocaltext.__main__:main"

[tool.setuptools.packages.find]
# This tells setuptools that the packages are under the 'src' directory
where = ["src"]
